feat: add rust port (#8)
Add a rust port of melody and move the golang code into a separate
directory.

The rust port was scaffolded out by AI and manually cleaned up by me.
There's still room to test & improve this but it is at a functional
state. I think we can build on this