chore: use option value for cur_citation_byte_index (#12)
Refactor the datatype for `cur_citation_byte_index`. In the Go
implementation `cur_citation_byte_index` was a `int` because it used
`-1` to represent when no citation was being parsed. In Rust that
concept is better handled by `Option<usize>`. this makes the logic
cleaner and removes some of the typecasting between int and uint